Move primitive transaction types to separate module.
