+++ lib/core/cardano-wallet-core.cabal
@@ -164,6 +164,7 @@ library
       Cardano.Wallet.Primitive.Types.ChimericAccount
       Cardano.Wallet.Primitive.Types.Coin
       Cardano.Wallet.Primitive.Types.Hash
+      Cardano.Wallet.Primitive.Types.Tx
       Cardano.Wallet.Registry
       Cardano.Wallet.Transaction
       Cardano.Wallet.Unsafe
