+++ lib/core/src/Cardano/Wallet/Primitive/Types/Tx.hs
@@ -1,0 +1,338 @@
+{-# LANGUAGE BangPatterns #-}
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE DerivingStrategies #-}
+{-# LANGUAGE DuplicateRecordFields #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE OverloadedLabels #-}
+{-# LANGUAGE TypeApplications #-}
+
+module Cardano.Wallet.Primitive.Types.Tx
+    (
+    -- * Types
+      Tx (..)
+    , TxIn (..)
+    , TxOut (..)
+    , TxChange (..)
+    , TxMeta (..)
+    , TxStatus (..)
+    , SealedTx (..)
+    , UnsignedTx (..)
+    , TransactionInfo (..)
+    , Direction (..)
+
+    -- * Functions
+    , fromTransactionInfo
+    , inputs
+    , isPending
+    , toTxHistory
+    , txIns
+    , txMetadataIsNull
+
+    ) where
+
+import Prelude
+
+import Cardano.Api.Typed
+    ( TxMetadata (..) )
+import Cardano.Slotting.Slot
+    ( SlotNo (..) )
+import Cardano.Wallet.Orphans
+    ()
+import Cardano.Wallet.Primitive.Types.Address
+    ( Address (..) )
+import Cardano.Wallet.Primitive.Types.ChimericAccount
+    ( ChimericAccount (..) )
+import Cardano.Wallet.Primitive.Types.Coin
+    ( Coin (..) )
+import Cardano.Wallet.Primitive.Types.Hash
+    ( Hash (..) )
+import Control.DeepSeq
+    ( NFData (..) )
+import Data.ByteArray
+    ( ByteArrayAccess )
+import Data.ByteString
+    ( ByteString )
+import Data.Generics.Internal.VL.Lens
+    ( view )
+import Data.Generics.Labels
+    ()
+import Data.List.NonEmpty
+    ( NonEmpty (..) )
+import Data.Quantity
+    ( Quantity (..) )
+import Data.Text.Class
+    ( CaseStyle (..)
+    , FromText (..)
+    , ToText (..)
+    , fromTextToBoundedEnum
+    , toTextFromBoundedEnum
+    )
+import Data.Map.Strict
+    ( Map )
+import Data.Set
+    ( Set )
+import Data.Time.Clock
+    ( UTCTime )
+import Data.Word
+    ( Word32 )
+import Fmt
+    ( Buildable (..)
+    , blockListF'
+    , fixedF
+    , nameF
+    , ordinalF
+    , prefixF
+    , suffixF
+    , tupleF
+    )
+import GHC.Generics
+    ( Generic )
+import Numeric.Natural
+    ( Natural )
+
+import qualified Data.Map.Strict as Map
+import qualified Data.Set as Set
+import qualified Data.Text.Lazy.Builder as Builder
+
+-- | Primitive @Tx@-type.
+--
+-- Currently tailored for jormungandr in that inputs are @(TxIn, Coin)@
+-- instead of @TxIn@. We might have to revisit this when supporting another
+-- node.
+data Tx = Tx
+    { txId
+        :: Hash "Tx"
+        -- ^ Jörmungandr computes transaction id by hashing the full content of
+        -- the transaction, which includes witnesses. Therefore, we need either
+        -- to keep track of the witnesses to be able to re-compute the tx id
+        -- every time, or, simply keep track of the id itself.
+    , resolvedInputs
+        :: ![(TxIn, Coin)]
+        -- ^ NOTE: Order of inputs matters in the transaction representation.
+        -- The transaction id is computed from the binary representation of a
+        -- tx, for which inputs are serialized in a specific order.
+    , outputs
+        :: ![TxOut]
+        -- ^ NOTE: Order of outputs matters in the transaction representations.
+        -- Outputs are used as inputs for next transactions which refer to them
+        -- using their indexes. It matters also for serialization.
+    , withdrawals
+        :: !(Map ChimericAccount Coin)
+        -- ^ Withdrawals (of funds from a registered reward account) embedded in
+        -- a transaction. The order does not matter.
+    , metadata
+        :: !(Maybe TxMetadata)
+        -- ^ Semi-structured application-specific extension data stored in the
+        -- transaction on chain.
+        --
+        -- This is not to be confused with 'TxMeta', which is information about
+        -- a transaction derived from the ledger.
+        --
+        -- See Appendix E of
+        -- <https://hydra.iohk.io/job/Cardano/cardano-ledger-specs/delegationDesignSpec/latest/download-by-type/doc-pdf/delegation_design_spec Shelley Ledger: Delegation/Incentives Design Spec>.
+    } deriving (Show, Generic, Ord, Eq)
+
+instance NFData Tx
+
+instance Buildable Tx where
+    build (Tx tid ins outs ws md) = mempty
+        <> build tid
+        <> build ("\n" :: String)
+        <> blockListF' "inputs" build (fst <$> ins)
+        <> blockListF' "outputs" build outs
+        <> blockListF' "withdrawals" tupleF (Map.toList ws)
+        <> nameF "metadata" (maybe "" build md)
+
+txIns :: Set Tx -> Set TxIn
+txIns = foldMap (Set.fromList . inputs)
+
+inputs :: Tx -> [TxIn]
+inputs = map fst . resolvedInputs
+
+data TxIn = TxIn
+    { inputId
+        :: !(Hash "Tx")
+    , inputIx
+        :: !Word32
+    } deriving (Show, Generic, Eq, Ord)
+
+instance NFData TxIn
+
+instance Buildable TxIn where
+    build txin = mempty
+        <> ordinalF (inputIx txin + 1)
+        <> " "
+        <> build (inputId txin)
+
+data TxOut = TxOut
+    { address
+        :: !Address
+    , coin
+        :: !Coin
+    } deriving (Show, Generic, Eq, Ord)
+
+data TxChange derivationPath = TxChange
+    { address
+        :: !Address
+    , amount
+        :: !Coin
+    , derivationPath
+        :: derivationPath
+    } deriving (Show, Generic, Eq, Ord)
+
+instance NFData TxOut
+
+instance Buildable TxOut where
+    build txout = mempty
+        <> build (coin txout)
+        <> " @ "
+        <> prefixF 8 addrF
+        <> "..."
+        <> suffixF 8 addrF
+      where
+        addrF = build $ view #address txout
+
+instance Buildable (TxIn, TxOut) where
+    build (txin, txout) = build txin <> " ==> " <> build txout
+
+-- | Additional information about a transaction, derived from the transaction
+-- and ledger state. This should not be confused with 'TxMetadata' which is
+-- application-specific data included with the transaction.
+data TxMeta = TxMeta
+    { status :: !TxStatus
+    , direction :: !Direction
+    , slotNo :: !SlotNo
+    , blockHeight :: !(Quantity "block" Word32)
+    , amount :: !(Quantity "lovelace" Natural)
+    , expiry :: !(Maybe SlotNo)
+      -- ^ The slot at which a pending transaction will no longer be accepted
+      -- into mempools.
+    } deriving (Show, Eq, Ord, Generic)
+
+instance NFData TxMeta
+
+instance Buildable TxMeta where
+    build (TxMeta s d sl (Quantity bh) (Quantity a) mex) = mempty
+        <> (case d of; Incoming -> "+"; Outgoing -> "-")
+        <> fixedF @Double 6 (fromIntegral a / 1e6)
+        <> " " <> build s
+        <> " since " <> build sl <> "#" <> build bh
+        <> maybe mempty (\ex -> " (expires slot " <> build ex <> ")") mex
+
+data TxStatus
+    = Pending
+        -- ^ Created, but not yet in a block.
+    | InLedger
+        -- ^ Has been found in a block.
+    | Expired
+        -- ^ Time to live (TTL) has passed.
+    deriving (Show, Eq, Ord, Bounded, Enum, Generic)
+
+instance NFData TxStatus
+
+instance Buildable TxStatus where
+    build = Builder.fromText . toTextFromBoundedEnum SpacedLowerCase
+
+instance FromText TxStatus where
+    fromText = fromTextToBoundedEnum SnakeLowerCase
+
+instance ToText TxStatus where
+    toText = toTextFromBoundedEnum SnakeLowerCase
+
+-- | An unsigned transaction.
+--
+-- See 'Tx' for a signed transaction.
+--
+data UnsignedTx input output change = UnsignedTx
+    { unsignedInputs
+        :: NonEmpty input
+        -- Inputs are *necessarily* non-empty because Cardano requires at least
+        -- one UTxO input per transaction to prevent replayable transactions.
+        -- (each UTxO being unique, including at least one UTxO in the
+        -- transaction body makes it seemingly unique).
+
+    , unsignedOutputs
+        :: [TxOut]
+        -- Unlike inputs, it is perfectly reasonable to have empty outputs. The
+        -- main scenario where this might occur is when constructing a
+        -- delegation for the sake of submitting a certificate. This type of
+        -- transaction does not typically include any target output and,
+        -- depending on which input(s) get selected to fuel the transaction, it
+        -- may or may not include a change output should its value be less than
+        -- the minimal UTxO value set by the network.
+    , unsignedChange
+        :: [change]
+    }
+    deriving (Eq, Show)
+
+-- | The effect of a @Transaction@ on the wallet balance.
+data Direction
+    = Outgoing -- ^ The wallet balance decreases.
+    | Incoming -- ^ The wallet balance increases or stays the same.
+    deriving (Show, Bounded, Enum, Eq, Ord, Generic)
+
+instance NFData Direction
+
+instance Buildable Direction where
+    build = Builder.fromText . toTextFromBoundedEnum SpacedLowerCase
+
+instance FromText Direction where
+    fromText = fromTextToBoundedEnum SnakeLowerCase
+
+instance ToText Direction where
+    toText = toTextFromBoundedEnum SnakeLowerCase
+
+-- | @SealedTx@ is a serialised transaction that is ready to be submitted
+-- to the node.
+newtype SealedTx = SealedTx { getSealedTx :: ByteString }
+    deriving stock (Show, Eq, Generic)
+    deriving newtype (ByteArrayAccess)
+
+-- | True if the given metadata refers to a pending transaction
+isPending :: TxMeta -> Bool
+isPending = (== Pending) . (status :: TxMeta -> TxStatus)
+
+-- | Full expanded and resolved information about a transaction, suitable for
+-- presentation to the user.
+data TransactionInfo = TransactionInfo
+    { txInfoId :: !(Hash "Tx")
+    -- ^ Transaction ID of this transaction
+    , txInfoInputs :: ![(TxIn, Coin, Maybe TxOut)]
+    -- ^ Transaction inputs and (maybe) corresponding outputs of the
+    -- source. Source information can only be provided for outgoing payments.
+    , txInfoOutputs :: ![TxOut]
+    -- ^ Payment destination.
+    , txInfoWithdrawals :: !(Map ChimericAccount Coin)
+    -- ^ Withdrawals on this transaction.
+    , txInfoMeta :: !TxMeta
+    -- ^ Other information calculated from the transaction.
+    , txInfoDepth :: Quantity "block" Natural
+    -- ^ Number of slots since the transaction slot.
+    , txInfoTime :: UTCTime
+    -- ^ Creation time of the block including this transaction.
+    , txInfoMetadata :: !(Maybe TxMetadata)
+    -- ^ Application-specific extension data.
+    } deriving (Generic, Show, Eq)
+
+instance NFData TransactionInfo
+
+-- | Reconstruct a transaction info from a transaction.
+fromTransactionInfo :: TransactionInfo -> Tx
+fromTransactionInfo info = Tx
+    { txId = txInfoId info
+    , resolvedInputs = (\(a,b,_) -> (a,b)) <$> txInfoInputs info
+    , outputs = txInfoOutputs info
+    , withdrawals = txInfoWithdrawals info
+    , metadata = txInfoMetadata info
+    }
+
+-- | Test whether the given metadata map is empty.
+txMetadataIsNull :: TxMetadata -> Bool
+txMetadataIsNull (TxMetadata md) = Map.null md
+
+-- | Drop time-specific information
+toTxHistory :: TransactionInfo -> (Tx, TxMeta)
+toTxHistory info =
+    (fromTransactionInfo info, txInfoMeta info)
